+++ InscryptionAPI/Card/CardManager.cs
@@ -24,6 +24,8 @@ public static class CardManager
    public class CardAltPortraits
     {
         public Sprite PixelAlternatePortrait = null;
+        public Sprite SteelTrapPortrait = null;
+        public Sprite BrokenShieldPortrait = null;
     }
     private static readonly ConditionalWeakTable<CardInfo, CardExt> CardExtensionProperties = new();
     private static readonly ConditionalWeakTable<CardInfo, CardAltPortraits> CardAlternatePortraits = new();
@@ -268,6 +270,14 @@ public static class CardManager
     public static Sprite PixelAlternatePortrait(this CardInfo card)
     {
         return card.GetAltPortraits().PixelAlternatePortrait;
+    }
+    public static Sprite SteelTrapPortrait(this CardInfo card)
+    {
+        return card.GetAltPortraits().SteelTrapPortrait;
+    }
+    public static Sprite BrokenShieldPortrait(this CardInfo card)
+    {
+        return card.GetAltPortraits().BrokenShieldPortrait;
     }
 
     private const string ERROR = "ERROR";